[symex] Move the test generators into goto-symex/testgen/
First of seven commits partitioning the flat src/goto-symex/ level into named
subsystem directories (docs/roadmap/goto-symex-partition-plan.md §2). ctest,
pytest and test_gen_guard are a leaf after the #7683 include repair: nothing
else in goto-symex includes them.

Pure move — the only content edits are #include paths.
