+++ src/esbmc/bmc.h
@@ -8,8 +8,8 @@
 #include <goto-symex/reachability_tree.h>
 #include <goto-symex/symex_target_equation.h>
 #include <goto-symex/witnesses.h>
-#include <goto-symex/pytest.h>
-#include <goto-symex/ctest.h>
+#include <goto-symex/testgen/pytest.h>
+#include <goto-symex/testgen/ctest.h>
 #include <langapi/language_ui.h>
 #include <list>
 #include <map>
