+++ src/goto-symex/testgen/ctest.cpp
@@ -1,6 +1,6 @@
-#include <goto-symex/ctest.h>
+#include <goto-symex/testgen/ctest.h>
 #include <goto-symex/slice.h>
-#include <goto-symex/test_gen_guard.h>
+#include <goto-symex/testgen/test_gen_guard.h>
 #include <goto-symex/witnesses.h>
 #include <ac_config.h>
 #include <langapi/mode.h>
